requirements.txt
# Основные зависимости для ФОТОТАБЛИЦЫ
Pillow>=10.0.0
pillow-heif>=0.16.0
python-docx>=0.8.11
geopy>=2.4.0
